Move single "util" class to `helper` package
